requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
dataclasses-json==0.6.7
ddgs==9.9.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
executing==2.2.1
fake-useragent==2.2.0
fastjsonschema==2.21.2
fqdn==1.5.1
frozenlist==1.8.0
groq==0.36.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
langchain==1.2.4
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-groq==1.1.1
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.1
lark==1.3.1
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mistune==3.1.4
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==1.26.4
orjson==3.11.4
ormsgpack==1.12.1
overrides==7.7.0
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
patsy==1.0.2
pexpect==4.9.0
pip==25.3
platformdirs==4.5.0
plotly==6.5.1
plotly-express==0.4.1
primp==0.15.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.29.0
scipy==1.16.3
Send2Trash==1.8.3
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
statsmodels==0.14.6
tenacity==8.5.0
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uuid_utils==0.12.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0